requirements.txt
Django>=1.8.0,<1.9.0
factory_boy>=2.3.1,<3.0.0
